Mcp/mcp-client/requirements.txt
fastmcp>=2.10.6
openai>=1.3.0
httpx>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# 语音功能依赖
sounddevice>=0.4.6
numpy>=1.21.0
wave 